python-backend/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Core LlamaIndex ecosystem
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-readers-file>=0.1.0
llama-index-embeddings-openai>=0.1.0

# Vector storage
chromadb>=0.4.0

# Agno framework for intelligent reasoning
agno>=0.1.0

# Configuration management
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Additional utilities
aiofiles>=23.2.1